% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hochschule.R
\name{check_hochschule}
\alias{check_hochschule}
\title{Prüft Hochschulnamen und Kürzel}
\usage{
check_hochschule(x)
}
\arguments{
\item{x}{Ein Character-Vektor mit Namen oder Kürzeln, die überprüft werden sollen.}
}
\value{
Ein logischer Vektor: \code{TRUE}, wenn der Eintrag einer bekannten Hochschule
entspricht, sonst \code{FALSE}.
}
\description{
Diese Funktion gleicht Eingaben mit der beigelegten SQL-Liste
(\code{hochschulen_namen_kuerzel.sql}) ab, die in \verb{inst/extdata/}
im Paket \pkg{SVCleanR} enthalten ist.
Es werden sowohl die ausgeschriebenen Namen als auch die Kürzel berücksichtigt.
Zusätzlich wird eine Übersicht mit \code{message} ausgegeben.
}
\details{
Hinweise:
\itemize{
\item Es wird gemeldet, wenn mehrere unterschiedliche Hochschulen im Vektor vorkommen.
\item Nicht gefundene Werte, leere Strings oder \code{NA} werden explizit aufgelistet.
\item Die aktuelle Referenzliste der gültigen Werte ist hier einsehbar:
\url{http://srv-data01:30080/hex/hex/-/blob/main/hochschulen_namen_kuerzel.sql?ref_type=heads}
}
}
\examples{
is_valid_hochschule(c("Universität Mannheim", "U Mannheim", ""))
is_valid_hochschule(c("Universität Mannheim", "Universität Heidelberg"))

}
