% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_organisation.R
\name{check_organisation}
\alias{check_organisation}
\title{Überprüft Organisationseinträge auf verschiedene Fehler}
\usage{
check_organisation(data, organisation_col = "organisation")
}
\arguments{
\item{data}{Dataframe mit einer Spalte für Organisationen.}

\item{organisation_col}{Name der Organisationsspalte als Zeichenkette (Standard: "organisation").}
}
\value{
Dataframe mit problematischen Einträgen, deren Zeilennummer, Originalwert und Art des Problems.
}
\description{
Diese Funktion prüft die Werte einer Organisationsspalte in einem Dataframe auf verschiedene
häufige Fehlerquellen. Dazu gehören unter anderem falsche oder fehlende Trennzeichen,
unerwünschte Sonderzeichen, zu lange Einträge und andere Auffälligkeiten.
}
\details{
Folgende Checks werden durchgeführt:
\itemize{
\item Falsche oder fehlende Trennung mit " ; " (Leerzeichen, Semikolon, Leerzeichen)
\item Enthält das Zeichen "|"
\item Semikolon ohne beidseitiges Leerzeichen
\item Semikolon nur mit links oder rechts Leerzeichen
\item Zu lange Einträge (> 1000 Zeichen)
\item Enthält das Zeichen ">"
}

Leere oder fehlende Werte werden ignoriert.
}
